requirements.txt
atlassian-python-api>=3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
